+++ Src/Library/Endpoint/Auxiliary/EndpointDefinition.cs
@@ -48,6 +48,9 @@ public sealed class EndpointDefinition(Type endpointType, Type requestDtoType, T
     public EndpointSummary? EndpointSummary { get; private set; }
     public List<string>? EndpointTags { get; private set; }
     public string? FormDataContentType { get; private set; }
+#if NET7_0_OR_GREATER
+    public IdempotencyOptions? IdempotencyOptions { get; private set; }
+#endif
     public string? OverriddenRoutePrefix { get; private set; }
     public List<string>? PreBuiltUserPolicies { get; private set; }
     public Action<AuthorizationPolicyBuilder>? PolicyBuilder { get; private set; }
@@ -254,8 +257,20 @@ public sealed class EndpointDefinition(Type endpointType, Type requestDtoType, T
         new TEndpointGroup().Action(this);
     }
 
+#if NET7_0_OR_GREATER
     /// <summary>
-    /// set endpoint configurations options using an endpoint builder action ///
+    /// specify idempotency requirements for this endpoint
+    /// </summary>
+    /// <param name="options">the idempotency options</param>
+    public void Idempotency(Action<IdempotencyOptions>? options = null)
+    {
+        IdempotencyOptions ??= new();
+        options?.Invoke(IdempotencyOptions);
+    }
+#endif
+
+    /// <summary>
+    /// set endpoint configurations options using an endpoint builder action
     /// </summary>
     /// <param name="builder">the builder for this endpoint</param>
     public void Options(Action<RouteHandlerBuilder> builder)